Add SSL support and integrate dbinit into entrypoint.sh
Add libressl to rootfs.

Add SSL config to postgres templates.

Integrate dbinit into entrypoint.sh. The entrypoint.sh script will now check
for the existance of a file named .dbinit in /data, if not found dbinit will
run.  Otherwise it will do nothing.  This is useful for non persistant
databaes instances used for testing since postgres will be initialized on a
directory in the container that will be destroyed when the container is
stopped.  An external volume still needs to be mounted to /data for the db
to persist.